Finalise new doc structure
