+++ docs/source/gallery.md
@@ -1,4 +1,4 @@
-# Examples
+# Gallery
 
 ## Realistic noise
