+++ docs/source/in-the-box.md
@@ -2,7 +2,6 @@
 
 <!-- Introduction is in the module __init__.py file -->
 
-```{autodoc2-docstring} pysmo
-render_plugin = "myst"
-no_index = true
+```{eval-rst}
+.. automodule:: pysmo
 ```
