+++ docs/source/tools.md
@@ -2,10 +2,6 @@
 
 TODO: add text to __init__.py
 
-```{autodoc2-docstring} pysmo.tools
-render_plugin = "myst"
-```
-
 ## Subpackages
 
 TODO: list tools
