+++ pysmo/__init__.py
@@ -6,22 +6,40 @@ stored and how they are processed. This makes it easy to switch between data sou
 multiple different ones simultaneously, as well as integrate with existing workflows. The
 pysmo package consists of:
 
-- [Pysmo types](<project:types.md#Types>): these are *protocol classes*, which define the
+- :ref:`Pysmo types<types:Types>`: these are *protocol classes*, which define the
   types used in pysmo. They are intentially kept as simple as possible, and are
   derived from the types of data seismologists work with, rather than mirroring how
   data are stored in files.
-- [Functions](<project:functions.md#Functions>): pysmo includes functions that perform
+- :ref:`Compatible classes<classes:Classes>`: while pysmo doesn't have a particlular
+  *native* (file)format, it does rely on the classes holding seismological data to
+  be compatible with the types defined by pysmo. Pysmo includes a minimal implementation
+  of such a class for each type. They include additional methods or read-only attributes
+  for the pysmo types. These minimal classes can be used as base classes when adopting
+  existing classes to work with pysmo types, so that they may provide the same methods.
+- :ref:`Functions<functions:Functions>`: pysmo includes functions that perform
   common operations on these data types. They are meant to be building blocks
   that can be used to construct more complex processing algorithms.
-- [Tools](<project:tools.md#Tools>): we place functions here that are too complex to be
+- :ref:`Tools<tools:tools>`: we place functions here that are too complex to be
   considered basic operations, make extensive use of other functions, or can be grouped
   together under a particular topic.
 """
 
 from importlib.metadata import version
-from pysmo.core.sac.sacio import _SacIO  # noqa: F401
-from pysmo.core.sac.sac import SAC  # noqa: F401
-from pysmo.core.protocols import *  # noqa: F401,F403
-from pysmo.core.functions import *  # noqa: F401,F403
+from pysmo.core.types import (
+    Epicenter,
+    Hypocenter,
+    Seismogram,
+    Station,
+)
+# from pysmo.core.sac.sacio import _SacIO  # noqa: F401
+# from pysmo.core.sac.sac import SAC  # noqa: F401
+# from pysmo.core.functions import *  # noqa: F401,F403
 
 __version__ = version('pysmo')
+
+__all__ = [
+    'Epicenter',
+    'Hypocenter',
+    'Seismogram',
+    'Station',
+]
