+++ pysmo/core/sac/__init__.py
@@ -1,0 +1,6 @@
+from .sac import SAC
+from .sacio import _SacIO  # noqa: F401
+
+__all__ = [
+    'SAC',
+]
