+++ pysmo/functions/__init__.py
@@ -1,0 +1,5 @@
+from ..core.functions.detrend import detrend  # noqa: F401
+from ..core.functions.resample import resample  # noqa: F401
+from ..core.functions.filter_gauss import envelope, gauss  # noqa: F401
+from ..core.functions.plotseis import plotseis  # noqa: F401
+from ..core.functions.azdist import azimuth, backazimuth, distance  # noqa: F401
